hello_unstable: random-readdir-order filesystem
This filesystem intentionally randomizes the results of the Readdir call on
every invocation; this is useful to show that de-duplicating the file descriptor
for opendir/readdir/releasedir is not a good idea, at least when mixed with
remotes that can not guarantee a sort order.  Theoretically we can sort it
ourselves, but this still implies that any sort of add/delete will be
problematic.

Reproduce the issue by building this filesystem, mounting it at a place like
"my_mount", and run something like this:

    for i in $(seq 1 10); do
      python -c 'import os; print len(set(os.listdir("my_mount")))' &
    done

In my tests, these python calls always receive 500 entries like they should, but
usually when there were multiple readdirs happening at once (due to the python
processes being backgrounded), most of the readdirs received duplicate entries.
In our production filesystem, this actually ends up missing entires
occasionally, I assume because we end up in the middle of an entry and return an
empty string, which we don't do in this case since our filenames are of a
certain length (or *not* of a certain length).